Update package.json and package-lock.json dependencies
This commit includes the upgrade of multiple dependencies versions in both the package.json and package-lock.json files. Importantly, the "@kuzrwkd/skys-core" and "playwright-core" base packages as well as several devDependencies such as "@typescript-eslint/parser", "esbuild", and "eslint" have been updated. This is essential for ensuring that the codebase is utilizing the most recent